Osa4

+++ Osa4/blogilista/controllers/blogs.js
@@ -1,32 +1,48 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+    return authorization.substring(7)
+  }
+  return null
+}
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({})
-    .populate('user', {username: 1, name: 1})
+    .populate('user', { username: 1, name: 1 })
   response.json(blogs.map(blog => blog.toJSON()))
 })
+
 
 blogsRouter.post('/', async (request, response, next) => {
   const body = request.body
 
-  //const user = await User.findById(body.userId)
-  const user = await User.findOne({})
-  body.user = user['_id']
-
-  if (!('title' in body) || !('url' in body)) {
-    return response.status(400).json({ error: 'Bad request' })
-  }
-
-  const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes,
-    user: user._id
-  })
   try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    //const user = await User.findById(body.userId)
+    const user = await User.findOne({})
+    body.user = user['_id']
+
+    if (!('title' in body) || !('url' in body)) {
+      return response.status(400).json({ error: 'Bad request' })
+    }
+
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes,
+      user: user._id
+    })
+
     const savedBlog = await blog.save()
     user.blogs = user.blogs.concat(savedBlog.id)
     await user.save()
@@ -51,6 +67,17 @@ blogsRouter.get('/:id', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const b = await Blog.findById(request.params.id)
+    if (decodedToken.id !== b.user.toString()) {
+      return response.status(401).json({ error: 'you can not delete someones elses blog' })
+    }
+    //const b = await Blog.findById(request.params.id)
+    //const userId = blog.user.toString()
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
   } catch (exception) {
